Add files via upload
First task 2-1 is to design and implement a class to describe the circle by the coordinates of the center and the length of the radius with functional capabilities to enter the length of the radius and the coordinates of the center, calculate the length of the circle and the face of the circle. In the main program, an object of the class is declared and information about it is displayed.