Add files via upload
To design and implement a class to set a triangle by long side lengths with capabilities to enter the length of the page, to double the side of the triangle, calculate the perimeter and face of the triangle, and print the page image of the triangle, the perimeter and the face. In the main program, you must declare the object's triangle face, enter values ​​for its side, enter a triangle with twice the page dimensions, and print the page size, perimeters, and both triangles.